+++ client/src/actions/authActions.jsx
@@ -103,6 +103,7 @@ export function authError(error) {
 
 export function signoutUser() {
   localStorage.removeItem('token');
+  browserHistory.push('/')
   return { type: UNAUTH_USER };
 }
