+++ client/src/containers/Nav.jsx
@@ -42,7 +42,7 @@ class Nav extends Component {
         </div>
         <div className="header_menus">
           <div className="header__wrapper">
-            <Link className="header__findmentor" to="/find_mentor"><h4>Find Mentor</h4></Link>
+            <Link className="header__findmentor" to="/find_mentor"><h4>Find Mentors</h4></Link>
           </div>
         </div>
       </div>
